pyproject.toml
[tool.poetry]
name = "final-project"
version = "0.1.0"
description = ""
authors = ["rahimi_h@yaftyar.ir <rahimi_h>"]
readme = "README.md"
packages = [{include = "final_project"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
django = "^4.1.4"
psycopg2-binary = "^2.9.5"
pip = "^22.3.1"
wheel = "^0.38.4"
setuptools = "^65.6.3"
pillow = "^9.3.0"
django-admin-relation-links = "^0.2.5"
django-cron = "^0.6.0"
django-crontab = "^0.7.1"
django-object-actions = "^4.1.0"
django-debug-toolbar = "^3.8.1"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"
dj-database-url = "^1.2.0"
python-decouple = "^3.7"
django-admin-interface = "^0.24.2"
django-thumbnails = "^0.7.0"
django-redis = "^5.2.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
httpie = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
